ui improvements

+++ app/src/main/java/com/f2h/f2h_buyer/screens/group/pre_order/PreOrderModel.kt
@@ -8,11 +8,11 @@ data class PreOrderModel (
     var itemName: String = "",
     var itemDescription: String = "",
     var farmerName: String = "",
-    var price: Float = (0).toFloat(),
+    var price: Float = 0F,
     var itemUom: String = "",
     var orderId: Long = 0,
-    var orderedQuantity: Float = (0).toFloat(),
+    var orderedQuantity: Float = 0F,
     var orderUom: String = "",
-    var orderAmount: Float = (0).toFloat(),
-    var discountAmount: Float = (0).toFloat()
+    var orderAmount: Float = 0F,
+    var discountAmount: Float = 0F
 )
